refactor(ios): split oversized SwiftUI views and view models (#348)
UI-behavior-preserving structural extraction across the view layer; no
suppressions:

- ActiveWorkoutView split into +ExerciseList / +Actions (set-action preamble
  and rest-timer/Live-Activity blocks deduplicated, order preserved).
- ActiveWorkoutViewModel.buildDisplayItems decomposed (c15 -> helpers);
  collapse-state queries take a CollapseContext struct (test call sites
  migrated mechanically).
- SetRowView (983 lines) split per state into +CurrentSet / +CompletedRow /
  +DropSet alongside the existing +Weight companion.
- WorkoutDetailView's plan-grouping logic moved to a PlanDisplayBuilder enum
  (c14 -> all <=10, duplicated superset handling unified); WorkoutsView,
  HistoryDetailView, HomeView, WorkoutSummaryView gained subview companion
  files; InboxSectionView's row became InboxRowView; EditPlanExerciseSheet's
  save path deduplicated against parseMarkdownIntoForm (was a line-for-line
  copy).
- SettingsView's iPad detail switch routed through per-section builders;
  AuthenticationStore's refresh flow moved verbatim to +Refresh;
  SessionStore's >5-param functions use defaulted params (call sites
  untouched).

Accessibility identifiers verified unchanged (multiset comparison per file
group). Baseline regenerated: 50 -> 21 entries, removals only — only the
CloudKit sync services (#346) remain. Smoke plan: 914 unit + smoke UI
tests, 0 failures.

Part of #351. Closes #348.
